YUK-17919-Duplicate Data Pruning go live (2)

+++ yukon-client/yukon-web/src/com/cannontech/web/maintenance/tasks/ScheduledRphDuplicateDeletionExecutionTask.java
@@ -3,6 +3,8 @@ package com.cannontech.web.maintenance.tasks;
 import org.joda.time.Instant;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import com.cannontech.common.config.ConfigurationSource;
+import com.cannontech.common.config.MasterConfigBoolean;
 import com.cannontech.common.events.loggers.SystemEventLogService;
 import com.cannontech.core.dao.DatabaseManagementDao;
 import com.cannontech.jobs.support.YukonTaskBase;
@@ -10,13 +12,17 @@ import com.cannontech.jobs.support.YukonTaskBase;
 public class ScheduledRphDuplicateDeletionExecutionTask extends YukonTaskBase {
     @Autowired private DatabaseManagementDao databaseManagementDao;
     @Autowired private SystemEventLogService systemEventLogService;
+    @Autowired private ConfigurationSource configurationSource;
 
     @Override
     public void start() {
-        // This is done just to inactive Duplicate Deletion Task
-        /*Instant start = new Instant();
-        int numDeleted = databaseManagementDao.deleteRphDuplicates();
-        Instant finish = new Instant();
-        systemEventLogService.rphDeleteDuplicates(numDeleted, start, finish);*/
+        // TODO This is done just to inactive Duplicate Deletion Task (YUK-17919).
+        boolean devMode = configurationSource.getBoolean(MasterConfigBoolean.DEVELOPMENT_MODE);
+        if (devMode) {
+            Instant start = new Instant();
+            int numDeleted = databaseManagementDao.deleteRphDuplicates();
+            Instant finish = new Instant();
+            systemEventLogService.rphDeleteDuplicates(numDeleted, start, finish);
+        }
     }
 }
